requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23

# Vector Store and Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
torch>=2.2.0
torchvision>=0.17.0

# LLM Inference
llama-cpp-python==0.2.90

# OCR and Image Processing
pytesseract==0.3.10
easyocr==1.7.1
opencv-python==4.8.1.78
Pillow==10.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Text Processing
nltk==3.8.1
numpy==1.26.2

# Visualization
matplotlib==3.8.2

# Security
bcrypt==4.1.1

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Scheduling
APScheduler==3.10.4

# Voice Input/Output
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1

# Handwriting Recognition
google-cloud-vision==3.4.5
opencv-contrib-python==4.8.1.78

# Utilities
requests==2.31.0
tqdm==4.66.1
